# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Score API",
        "contact": {
            "name": "Score API Developer Team",
            "url": "https://github.com/pine09/ScoreAPI"
        },
        "version": "1.0"
    },
    "host": "score.app",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/admin/mahasiswa": {
            "get": {
                "tags": [
                    "Admin on Mahasiswa"
                ],
                "summary": "Mengambil semua data mahasiswa yang ada.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data mahasiswa. (max 5 per page)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mahasiswa"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin on Mahasiswa"
                ],
                "summary": "Mendaftarkan mahasiswa baru.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data Mahasiswa Baru.",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mahasiswa"
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data mahasiswa baru."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "409": {
                        "description": "Conflict."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    }
                }
            }
        },
        "/api/v1/admin/mahasiswa/{id}": {
            "get": {
                "tags": [
                    "Admin on Mahasiswa"
                ],
                "summary": "Mengambil data mahasiswa spesifik",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data mahasiswa yang dicari"
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin on Mahasiswa"
                ],
                "summary": "Mengedit data diri mahasiswa.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data Mahasiswa (Edit).",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mahasiswa"
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data mahasiswa edited."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    },
                    "409": {
                        "description": "Conflict."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin on Mahasiswa"
                ],
                "summary": "Men-delete data mahasiswa.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data mahasiswa deleted."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mahasiswa": {
            "properties": {
                "NIM": {
                    "type": "string"
                },
                "foto": {
                    "description": "path to image. nullable field",
                    "type": "string"
                },
                "nama_depan": {
                    "type": "string"
                },
                "nama_belakang": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "alamat": {
                    "type": "text"
                },
                "jenis_kelamin": {
                    "type": "enum('Laki-Laki', 'Perempuan')"
                },
                "jurusan_id": {
                    "type": "integer",
                    "format": "int32"
                },
                "konsentrasi_id": {
                    "description": "nullable field",
                    "type": "integer",
                    "format": "int32"
                },
                "angkatan": {
                    "type": "string"
                },
                "kelas": {
                    "type": "string"
                }
            }
        },
...
    }
}